Experimental Beta Version/Dr_Agent - With MistralAI/requirements.txt
python-telegram-bot==20.3
python-dotenv
requests
urllib3
backoff
